FastAPI/ObstacleDetection/requirements.txt
"""Requirements for Obstacle Detection service."""

fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
pillow>=10.0.0
huggingface-hub>=0.17.0
